Move continued programs from community to current offerings page
The Aura for Life and Teachers Training programs are now displayed
under the Current Offerings page instead of the Community page,
keeping all program offerings in one place.
