+++ src/app/(site)/community/CommunityClient.tsx
@@ -1,32 +1,14 @@
 'use client';
 
-import { useRef, useState } from 'react';
-import { motion, useInView, AnimatePresence } from 'framer-motion';
-import { ChevronDown } from 'lucide-react';
-import Link from 'next/link';
+import { useRef } from 'react';
+import { motion, useInView } from 'framer-motion';
 import { cn } from '@/lib/utils';
-import { freePrograms, paidPrograms } from '@/lib/data';
+import { freePrograms } from '@/lib/data';
 import type { CommunityProgram } from '@/lib/data';
 import SubscribeCalendar from '@/components/ui/SubscribeCalendar';
 
-import type { TimeZoneEntry } from '@/lib/data/types';
 import PageHero from '@/components/sections/PageHero';
 import InvitationCTA from '@/components/sections/InvitationCTA';
-
-// =============================================================================
-// TIMEZONE SELECTOR (matches ScheduleTable on offerings page)
-// =============================================================================
-
-type TimezoneKey = keyof TimeZoneEntry;
-
-const timezoneLabels: Record<TimezoneKey, string> = {
-  sanJose: 'San Jose',
-  bogota: 'Bogota',
-  newYork: 'New York',
-  brasilia: 'Brasilia',
-  london: 'London',
-  madrid: 'Madrid',
-};
 
 // =============================================================================
 // EASE CONSTANT
@@ -42,17 +24,11 @@ function ActivityCard({
   program,
   index,
   inView,
-  expanded,
-  onToggle,
 }: {
   program: CommunityProgram;
   index: number;
   inView: boolean;
-  expanded?: boolean;
-  onToggle?: () => void;
 }) {
-  const hasDetails = program.scheduleCycles || program.investment;
-
   return (
     <motion.div
       initial={{ opacity: 0, y: 24 }}
@@ -65,8 +41,7 @@ function ActivityCard({
         'border border-rose-200/50 dark:border-rose-800/20',
         'hover:border-rose-300 dark:hover:border-rose-700/40',
         'hover:shadow-lg hover:shadow-rose-500/5',
-        'transition-all duration-500',
-        expanded && 'border-rose-300 dark:border-rose-700/40 shadow-lg shadow-rose-500/5'
+        'transition-all duration-500'
       )}
     >
       <h3 className="font-serif text-[clamp(1.15rem,2.5vw,1.5rem)] leading-tight tracking-tight mb-3 text-[var(--color-foreground)]">
@@ -160,27 +135,6 @@ function ActivityCard({
         </div>
       )}
 
-      {/* View Details toggle for programs with schedule/investment */}
-      {hasDetails && onToggle && (
-        <button
-          onClick={onToggle}
-          className={cn(
-            'mt-6 inline-flex items-center gap-2',
-            'text-sm font-medium',
-            'text-[var(--color-foreground-muted)]',
-            'hover:text-[var(--color-foreground)]',
-            'transition-colors duration-200'
-          )}
-        >
-          {expanded ? 'Hide Schedule & Investment' : 'View 2026 Schedule & Investment'}
-          <ChevronDown
-            className={cn(
-              'w-4 h-4 transition-transform duration-300',
-              expanded && 'rotate-180'
-            )}
-          />
-        </button>
-      )}
     </motion.div>
   );
 }
@@ -190,21 +144,11 @@ function ActivityCard({
 // =============================================================================
 
 export default function CommunityClient() {
-  const [expandedProgramId, setExpandedProgramId] = useState<string | null>(null);
-  const [timezone, setTimezone] = useState<TimezoneKey>('newYork');
-
   const visionRef = useRef<HTMLElement>(null);
   const visionInView = useInView(visionRef, { once: true, margin: '-100px' });
 
   const freeRef = useRef<HTMLElement>(null);
   const freeInView = useInView(freeRef, { once: true, margin: '-100px' });
-
-  const paidRef = useRef<HTMLElement>(null);
-  const paidInView = useInView(paidRef, { once: true, margin: '-100px' });
-
-  const handleToggle = (id: string) => {
-    setExpandedProgramId((prev) => (prev === id ? null : id));
-  };
 
   return (
     <>
@@ -304,242 +248,7 @@ export default function CommunityClient() {
         </div>
       </section>
 
-      {/* 4. Continued Programs */}
-      <section ref={paidRef} className="section-padding">
-        <div className="container-premium max-w-3xl mx-auto">
-          <motion.p
-            initial={{ opacity: 0, y: 20 }}
-            animate={paidInView ? { opacity: 1, y: 0 } : {}}
-            transition={{ duration: 0.6, ease }}
-            className="label-sacred mb-6"
-          >
-            Continued Programs
-          </motion.p>
-          <motion.h2
-            initial={{ opacity: 0, y: 24 }}
-            animate={paidInView ? { opacity: 1, y: 0 } : {}}
-            transition={{ duration: 0.6, delay: 0.1, ease }}
-            className="font-serif text-[clamp(1.5rem,3.5vw,2.5rem)] leading-tight tracking-tight mb-4"
-          >
-            Deepening the Path
-          </motion.h2>
-          <motion.p
-            initial={{ opacity: 0, y: 24 }}
-            animate={paidInView ? { opacity: 1, y: 0 } : {}}
-            transition={{ duration: 0.6, delay: 0.15, ease }}
-            className="text-base text-[var(--color-foreground-muted)] leading-relaxed mb-8"
-          >
-            For those ready to go further, these programs offer sustained
-            practice, mentorship, and the opportunity to embody and share this
-            work at a deeper level.
-          </motion.p>
-          <div className="space-y-4">
-            {paidPrograms.map((program, i) => {
-              const isExpanded = expandedProgramId === program.id;
-
-              return (
-                <div key={program.id}>
-                  <ActivityCard
-                    program={program}
-                    index={i}
-                    inView={paidInView}
-                    expanded={isExpanded}
-                    onToggle={() => handleToggle(program.id)}
-                  />
-
-                  {/* Expanded: Schedule + Investment */}
-                  <AnimatePresence>
-                    {isExpanded && (
-                      <motion.div
-                        initial={{ opacity: 0, height: 0 }}
-                        animate={{ opacity: 1, height: 'auto' }}
-                        exit={{ opacity: 0, height: 0 }}
-                        transition={{ duration: 0.5, ease }}
-                        className="overflow-hidden"
-                      >
-                        {/* Schedule section */}
-                        {program.scheduleCycles && (
-                          <div className="mt-8 mb-8">
-                            <div className="flex flex-wrap items-end justify-between gap-4 mb-6">
-                              <div>
-                                <p className="label-sacred mb-3">Schedule</p>
-                                <h3 className="font-serif text-[clamp(1.25rem,3vw,2rem)] leading-tight tracking-tight">
-                                  2026 Class Schedule
-                                </h3>
-                              </div>
-                              <div className="flex items-center">
-                                <label className="label-sacred mr-3" htmlFor={`tz-select-${program.id}`}>
-                                  Timezone
-                                </label>
-                                <div className="relative">
-                                  <select
-                                    id={`tz-select-${program.id}`}
-                                    value={timezone}
-                                    onChange={(e) => setTimezone(e.target.value as TimezoneKey)}
-                                    className={cn(
-                                      'appearance-none cursor-pointer',
-                                      'bg-[var(--color-background-subtle)] border border-[var(--color-border)]',
-                                      'rounded-lg px-4 py-2 pr-9',
-                                      'text-sm text-[var(--color-foreground-subtle)]',
-                                      'focus-premium',
-                                      'transition-colors duration-200'
-                                    )}
-                                  >
-                                    {(Object.keys(timezoneLabels) as TimezoneKey[]).map((tz) => (
-                                      <option key={tz} value={tz}>
-                                        {timezoneLabels[tz]}
-                                      </option>
-                                    ))}
-                                  </select>
-                                  <ChevronDown
-                                    className="absolute right-2.5 top-1/2 -translate-y-1/2 w-4 h-4 text-[var(--color-foreground-faint)] pointer-events-none"
-                                    aria-hidden="true"
-                                  />
-                                </div>
-                              </div>
-                            </div>
-
-                            <div className="space-y-4">
-                              {program.scheduleCycles.map((cycle) => (
-                                <div
-                                  key={cycle.id}
-                                  className={cn(
-                                    'border border-[var(--color-border)] rounded-xl overflow-hidden',
-                                    'bg-[var(--color-background-elevated)]',
-                                    'shadow-[var(--shadow-md)]'
-                                  )}
-                                >
-                                  {/* Cycle header */}
-                                  <div className="px-4 py-3 sm:px-5 sm:py-4 md:px-6 md:py-5">
-                                    <h4 className="font-serif text-base sm:text-lg md:text-xl text-[var(--color-foreground)] tracking-tight">
-                                      {cycle.title}
-                                    </h4>
-                                  </div>
-
-                                  {/* Month groups */}
-                                  <div className="px-4 pb-4 sm:px-5 sm:pb-5 md:px-6 md:pb-6">
-                                    {cycle.months.map((month, monthIdx) => (
-                                      <div key={month.month} className={cn(monthIdx > 0 && 'mt-4')}>
-                                        <p className="label-sacred mb-2 text-xs">
-                                          {month.month}
-                                        </p>
-                                        <div className="space-y-1">
-                                          {month.sessions.map((session, sIdx) => (
-                                            <div
-                                              key={sIdx}
-                                              className={cn(
-                                                'flex flex-col sm:flex-row sm:items-baseline sm:justify-between gap-0.5 sm:gap-3 py-2.5 sm:py-2 text-sm',
-                                                sIdx < month.sessions.length - 1 &&
-                                                  'border-b border-[var(--color-border-subtle)]'
-                                              )}
-                                            >
-                                              <span className="text-[var(--color-foreground-subtle)] font-medium">
-                                                {session.date}
-                                              </span>
-                                              <span className="text-[var(--color-foreground-muted)] tabular-nums shrink-0 text-xs sm:text-sm">
-                                                {session.time[timezone]}
-                                              </span>
-                                            </div>
-                                          ))}
-                                        </div>
-                                      </div>
-                                    ))}
-                                  </div>
-                                </div>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-
-                        {/* Investment section */}
-                        {program.investment && (
-                          <div className="pt-6 pb-4">
-                            <p className="label-sacred mb-3">Investment</p>
-                            <h3 className="font-serif text-[clamp(1.25rem,3vw,2rem)] leading-tight tracking-tight mb-6">
-                              {program.title}
-                            </h3>
-
-                            <div className={cn(
-                              'grid gap-4',
-                              program.investment.length === 3
-                                ? 'sm:grid-cols-3'
-                                : 'sm:grid-cols-2'
-                            )}>
-                              {program.investment.map((option) => (
-                                <div
-                                  key={option.id}
-                                  className={cn(
-                                    'rounded-xl p-5',
-                                    'border border-[var(--color-border)]',
-                                    'bg-[var(--color-background-elevated)]',
-                                    'shadow-[var(--shadow-md)]',
-                                    'text-center'
-                                  )}
-                                >
-                                  <p className="font-serif text-lg text-[var(--color-foreground)] mb-1">
-                                    {option.label}
-                                  </p>
-                                  <p className="text-xs text-[var(--color-foreground-faint)] mb-3">
-                                    {option.period}
-                                  </p>
-                                  <p className="font-serif text-2xl text-[var(--color-foreground)] tracking-tight">
-                                    {option.price}
-                                  </p>
-                                </div>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-                      </motion.div>
-                    )}
-                  </AnimatePresence>
-                </div>
-              );
-            })}
-          </div>
-
-          {/* CTA to Offerings */}
-          <motion.div
-            initial={{ opacity: 0, y: 16 }}
-            animate={paidInView ? { opacity: 1, y: 0 } : {}}
-            transition={{ duration: 0.5, delay: 0.4, ease }}
-            className="text-center mt-12"
-          >
-            <p className="text-base text-[var(--color-foreground-muted)] leading-relaxed mb-6 max-w-lg mx-auto">
-              Looking for our foundational courses? Explore The Rose + Aura 1
-              and Aura Reading Level 2 on our offerings page.
-            </p>
-            <Link
-              href="/offerings"
-              className={cn(
-                'inline-flex items-center gap-2 px-8 py-3.5 rounded-full',
-                'border-2 border-[var(--color-rose-clay)] text-[var(--color-rose-clay)]',
-                'text-sm font-medium',
-                'hover:bg-[var(--color-rose-clay)] hover:text-[var(--color-foreground-on-rose)]',
-                'transition-all duration-200',
-                'shadow-sm hover:shadow-md'
-              )}
-            >
-              View Current Offerings
-              <svg
-                className="w-4 h-4"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-                strokeWidth={2}
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M17 8l4 4m0 0l-4 4m4-4H3"
-                />
-              </svg>
-            </Link>
-          </motion.div>
-        </div>
-      </section>
-
-      {/* 5. Invitation CTA */}
+      {/* 4. Invitation CTA */}
       <InvitationCTA />
     </>
   );
